# backend/src/main/resources/api-docs.yml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/notes/{noteId}":{"get":{"tags":["note-controller"],"description":"Get detailed information about note including its content","operationId":"get","parameters":[{"name":"noteId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoteSerializer"}}}}}},"put":{"tags":["note-controller"],"description":"Update note: title, blocks, categories etc.","operationId":"update","parameters":[{"name":"noteId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteSerializer"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoteEditResponse"}}}}}},"delete":{"tags":["note-controller"],"description":"Delete a note","operationId":"delete","parameters":[{"name":"noteId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoteDeleteResponse"}}}}}}},"/api/v1/category/{categoryId}":{"put":{"tags":["category-controller"],"description":"Edit category, only name","operationId":"edit","parameters":[{"name":"categoryId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryEditRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryEditResponse"}}}}}},"delete":{"tags":["category-controller"],"description":"Delete category","operationId":"delete_1","parameters":[{"name":"categoryId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryEditResponse"}}}}}}},"/api/v1/notes":{"get":{"tags":["note-controller"],"description":"Get brief information about notes with pagination","operationId":"list","parameters":[{"name":"page","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoteBriefResponse"}}}}}},"post":{"tags":["note-controller"],"description":"Save a note","operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteSerializer"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoteEditResponse"}}}}}}},"/api/v1/category":{"get":{"tags":["category-controller"],"description":"Get a list of categories, briefly serialized, only id and name","operationId":"list_1","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryListResponse"}}}}}},"post":{"tags":["category-controller"],"description":"Create a category, accept category name","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryEditRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryEditResponse"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"description":"Registration","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"description":"Authentication","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/v1/dictionary/{word}":{"get":{"tags":["dictionary-controller"],"description":"Get a list of definitions of the word in English","operationId":"getWordDefinitionList","parameters":[{"name":"word","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GenericErrorResponse"},{"$ref":"#/components/schemas/FieldsValidationErrorResponse"}]}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"502":{"description":"Bad Gateway","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DictionaryDefinitionListResponse"}}}}}}}},"components":{"schemas":{"GenericErrorResponse":{"type":"object","properties":{"detail":{"type":"string"}}},"FieldsValidationErrorResponse":{"type":"object","properties":{"detail":{"type":"string"},"fieldErrors":{"type":"object","additionalProperties":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}}},"NoteBlockSerializer":{"required":["type"],"type":"object","properties":{"type":{"type":"string"}},"discriminator":{"propertyName":"type"}},"NoteChecklistBlockDataSerializer":{"required":["items"],"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/NoteChecklistBlockItemSerializer"}}}},"NoteChecklistBlockItemSerializer":{"required":["checked","text"],"type":"object","properties":{"text":{"type":"string"},"checked":{"type":"boolean"}}},"NoteChecklistBlockSerializer":{"required":["data"],"type":"object","allOf":[{"$ref":"#/components/schemas/NoteBlockSerializer"},{"type":"object","properties":{"id":{"type":"string"},"data":{"$ref":"#/components/schemas/NoteChecklistBlockDataSerializer"}}}]},"NoteCodeBlockDataSerializer":{"required":["code"],"type":"object","properties":{"code":{"type":"string"}}},"NoteCodeBlockSerializer":{"required":["data"],"type":"object","allOf":[{"$ref":"#/components/schemas/NoteBlockSerializer"},{"type":"object","properties":{"id":{"type":"string"},"data":{"$ref":"#/components/schemas/NoteCodeBlockDataSerializer"}}}]},"NoteContentSerializer":{"required":["blocks","version"],"type":"object","properties":{"version":{"type":"string"},"time":{"type":"integer","format":"int64"},"blocks":{"type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/NoteChecklistBlockSerializer"},{"$ref":"#/components/schemas/NoteCodeBlockSerializer"},{"$ref":"#/components/schemas/NoteHeaderBlockSerializer"},{"$ref":"#/components/schemas/NoteListBlockSerializer"},{"$ref":"#/components/schemas/NoteParagraphBlockSerializer"},{"$ref":"#/components/schemas/NoteQuoteBlockSerializer"}]}}}},"NoteHeaderBlockDataSerializer":{"required":["level","text"],"type":"object","properties":{"text":{"type":"string"},"level":{"type":"integer","format":"int32"}}},"NoteHeaderBlockSerializer":{"required":["data"],"type":"object","allOf":[{"$ref":"#/components/schemas/NoteBlockSerializer"},{"type":"object","properties":{"id":{"type":"string"},"data":{"$ref":"#/components/schemas/NoteHeaderBlockDataSerializer"}}}]},"NoteListBlockDataSerializer":{"required":["items","style"],"type":"object","properties":{"style":{"type":"string"},"items":{"type":"array","items":{"type":"string"}}}},"NoteListBlockSerializer":{"required":["data"],"type":"object","allOf":[{"$ref":"#/components/schemas/NoteBlockSerializer"},{"type":"object","properties":{"id":{"type":"string"},"data":{"$ref":"#/components/schemas/NoteListBlockDataSerializer"}}}]},"NoteParagraphBlockDataSerializer":{"required":["text"],"type":"object","properties":{"text":{"type":"string"}}},"NoteParagraphBlockSerializer":{"required":["data"],"type":"object","allOf":[{"$ref":"#/components/schemas/NoteBlockSerializer"},{"type":"object","properties":{"id":{"type":"string"},"data":{"$ref":"#/components/schemas/NoteParagraphBlockDataSerializer"}}}]},"NoteQuoteBlockDataSerializer":{"required":["alignment","caption","text"],"type":"object","properties":{"text":{"type":"string"},"caption":{"type":"string"},"alignment":{"type":"string"}}},"NoteQuoteBlockSerializer":{"required":["data"],"type":"object","allOf":[{"$ref":"#/components/schemas/NoteBlockSerializer"},{"type":"object","properties":{"id":{"type":"string"},"data":{"$ref":"#/components/schemas/NoteQuoteBlockDataSerializer"}}}]},"NoteSerializer":{"required":["categoryIds","content","title"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"content":{"$ref":"#/components/schemas/NoteContentSerializer"},"categoryIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"NoteEditResponse":{"type":"object","properties":{"detail":{"type":"string"},"note":{"$ref":"#/components/schemas/NoteEditSerializer"}}},"NoteEditSerializer":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CategoryEditRequest":{"required":["category"],"type":"object","properties":{"category":{"$ref":"#/components/schemas/CategorySerializer"}}},"CategorySerializer":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"maxLength":255,"minLength":1,"type":"string"}}},"CategoryEditResponse":{"type":"object","properties":{"detail":{"type":"string"},"category":{"$ref":"#/components/schemas/CategorySerializer"}}},"RegistrationRequest":{"required":["fullName","password","passwordRepeat","username"],"type":"object","properties":{"fullName":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"passwordRepeat":{"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"},"user":{"$ref":"#/components/schemas/UserDetailsSerializer"}}},"UserDetailsSerializer":{"type":"object","properties":{"username":{"type":"string"},"fullName":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"NoteBriefResponse":{"type":"object","properties":{"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"notes":{"type":"array","items":{"$ref":"#/components/schemas/NoteBriefSerializer"}}}},"NoteBriefSerializer":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"categoryIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"Definition":{"type":"object","properties":{"definition":{"type":"string"},"example":{"type":"string"},"synonyms":{"type":"array","items":{"type":"string"}},"antonyms":{"type":"array","items":{"type":"string"}}}},"DictionaryDefinitionListResponse":{"type":"object","properties":{"definitions":{"type":"array","items":{"$ref":"#/components/schemas/Definition"}}}},"CategoryListResponse":{"type":"object","properties":{"categories":{"type":"array","items":{"$ref":"#/components/schemas/CategorySerializer"}}}},"NoteDeleteResponse":{"type":"object","properties":{"detail":{"type":"string"},"id":{"type":"integer","format":"int64"}}}}}}